refactor: juice
